refactor: Data와 Domain Mapper 분리
