+++ app/src/main/java/woowacourse/shopping/data/mapper/ProductDtoMapper.kt
@@ -1,14 +1,12 @@
-package woowacourse.shopping.utils.mapper
+package woowacourse.shopping.data.mapper
 
-import woowacourse.shopping.data.model.RecentlyProductEntity
 import woowacourse.shopping.data.remote.dto.product.ProductDto
 import woowacourse.shopping.data.remote.dto.product.ProductResponse
 import woowacourse.shopping.domain.model.cart.CartItemCounter
 import woowacourse.shopping.domain.model.product.Product
-import woowacourse.shopping.domain.model.product.RecentlyProduct
 import woowacourse.shopping.domain.model.selector.ItemSelector
 
-object ProductMapper {
+object ProductDtoMapper {
     fun ProductResponse.toProducts(): List<Product> {
         return productDto.map { it.toProduct() }
     }
@@ -22,25 +20,6 @@ object ProductMapper {
             category = category,
             cartItemCounter = CartItemCounter(quantity),
             itemSelector = ItemSelector(),
-        )
-    }
-
-    fun RecentlyProductEntity.toRecentlyProduct(): RecentlyProduct {
-        return RecentlyProduct(
-            id = id,
-            productId = productId,
-            imageUrl = imageUrl,
-            name = name,
-            category = category,
-        )
-    }
-
-    fun RecentlyProduct.toRecentlyProductEntity(): RecentlyProductEntity {
-        return RecentlyProductEntity(
-            productId = productId,
-            name = name,
-            imageUrl = imageUrl,
-            category = category,
         )
     }
 }
