+++ app/src/main/java/woowacourse/shopping/data/remote/source/RemoteCouponDataSourceImpl.kt
@@ -4,7 +4,7 @@ import woowacourse.shopping.data.remote.api.CouponApiService
 import woowacourse.shopping.data.remote.api.NetworkManager
 import woowacourse.shopping.data.source.CouponDataSource
 import woowacourse.shopping.domain.model.coupon.Coupon
-import woowacourse.shopping.utils.mapper.CouponMapper.toCoupon
+import woowacourse.shopping.data.mapper.CouponMapper.toCoupon
 
 class RemoteCouponDataSourceImpl(
     private val couponService: CouponApiService = NetworkManager.couponService(),
