+++ app/src/main/java/woowacourse/shopping/data/repository/RecentlyProductRepositoryImpl.kt
@@ -2,11 +2,12 @@ package woowacourse.shopping.data.repository
 
 import android.content.Context
 import woowacourse.shopping.data.db.recently.RecentlyProductDatabase
+import woowacourse.shopping.domain.mapper.ProductEntityMapper.toRecentlyProduct
+import woowacourse.shopping.domain.mapper.ProductEntityMapper.toRecentlyProductEntity
 import woowacourse.shopping.domain.model.product.RecentlyProduct
 import woowacourse.shopping.domain.repository.RecentlyProductRepository
 import woowacourse.shopping.utils.exception.ErrorEvent
-import woowacourse.shopping.utils.mapper.ProductMapper.toRecentlyProduct
-import woowacourse.shopping.utils.mapper.ProductMapper.toRecentlyProductEntity
+
 
 class RecentlyProductRepositoryImpl(context: Context) : RecentlyProductRepository {
     private val recentlyProductDao =
